move events and engines into own packages
