+++ src/main/java/dev/aisandbox/launcher/SandboxServerCLIApplication.java
@@ -9,14 +9,9 @@ package dev.aisandbox.launcher;
 import dev.aisandbox.launcher.options.RuntimeUtils;
 import dev.aisandbox.server.engine.SimulationBuilder;
 import dev.aisandbox.server.engine.SimulationParameter;
-import dev.aisandbox.server.engine.SimulationRandomNumberGenerator;
 import dev.aisandbox.server.engine.SimulationRunner;
-import dev.aisandbox.server.engine.Theme;
 import dev.aisandbox.server.engine.exception.SimulationSetupException;
-import dev.aisandbox.server.engine.output.OutputRenderer;
 import dev.aisandbox.server.engine.setup.SimulationSettings;
-import dev.aisandbox.server.engine.telemetry.NullTelemetryEngine;
-import dev.aisandbox.server.engine.telemetry.TelemetryEngine;
 import java.io.IOException;
 import java.util.Arrays;
 import java.util.List;
