+++ src/main/java/dev/aisandbox/server/engine/setup/SimulationSettings.java
@@ -17,8 +17,8 @@ import dev.aisandbox.server.engine.network.NetworkAgent;
 import dev.aisandbox.server.engine.output.BitmapOutputRenderer;
 import dev.aisandbox.server.engine.output.NullOutputRenderer;
 import dev.aisandbox.server.engine.output.OutputRenderer;
-import dev.aisandbox.server.engine.telemetry.FileTelemetryEngine;
-import dev.aisandbox.server.engine.telemetry.NullTelemetryEngine;
+import dev.aisandbox.server.engine.telemetry.engine.FileTelemetryEngine;
+import dev.aisandbox.server.engine.telemetry.engine.NullTelemetryEngine;
 import dev.aisandbox.server.engine.telemetry.TelemetryEngine;
 import java.io.File;
 import java.util.ArrayList;
