--- src/main/java/dev/aisandbox/server/engine/telemetry/EpisodeLongScoreEvent.java
@@ -1,25 +1,0 @@
-/*
- * AI Sandbox - This program is free software: you can redistribute it and/or modify it under the
- * terms of version 3 of the GNU General Public License. See the README and LICENCE files for
- * more information.
- */
-
-package dev.aisandbox.server.engine.telemetry;
-
-import java.time.Instant;
-
-/**
- * Telemetry event to denote a simulation episode completing with a long integer score.
- *
- * @param simulationName      The name of the simulation
- * @param sessionID           The session identifier
- * @param episodeID           The episode identifier
- * @param timestamp The time the event was created
- * @param score               The score for the episode
- */
-public record EpisodeLongScoreEvent(String simulationName,
-                                    String sessionID,
-                                    String episodeID,
-                                    Instant timestamp,
-                                    long score) implements TelemetryEvent {
-}
