+++ src/main/java/dev/aisandbox/server/engine/telemetry/event/EpisodeAgentLongScoreEvent.java
@@ -1,11 +1,20 @@
 /*
+ *
+ *  * AI Sandbox - This program is free software: you can redistribute it and/or modify it under the
+ *  * terms of version 3 of the GNU General Public License. See the README and LICENCE files for
+ *  * more information.
+ *
+ */
+
+/*
  * AI Sandbox - This program is free software: you can redistribute it and/or modify it under the
  * terms of version 3 of the GNU General Public License. See the README and LICENCE files for
  * more information.
  */
 
-package dev.aisandbox.server.engine.telemetry;
+package dev.aisandbox.server.engine.telemetry.event;
 
+import dev.aisandbox.server.engine.telemetry.TelemetryEvent;
 import java.time.Instant;
 import java.util.List;
