+++ src/test/java/dev/aisandbox/server/simulation/bandit/TestRunBadBandit.java
@@ -15,7 +15,7 @@ import dev.aisandbox.server.engine.Theme;
 import dev.aisandbox.server.engine.exception.IllegalActionException;
 import dev.aisandbox.server.engine.output.NullOutputRenderer;
 import dev.aisandbox.server.engine.output.OutputRenderer;
-import dev.aisandbox.server.engine.telemetry.NullTelemetryEngine;
+import dev.aisandbox.server.engine.telemetry.engine.NullTelemetryEngine;
 import dev.aisandbox.server.simulation.bandit.model.BanditUpdateEnumeration;
 import java.util.Arrays;
 import java.util.List;
